refactor: seperate logic, add new viewModels
