+++ src/main/kotlin/components/MyText.kt
@@ -1,4 +1,4 @@
-package view
+package components
 
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
